📝 Updated documentation engine to typedoc
