+++ package.json
@@ -55,27 +55,25 @@
   },
   "homepage": "https://github.com/GV14982/async-airtable#readme",
   "dependencies": {
-    "node-fetch": "^2.6.1",
-    "typedoc": "^0.19.2"
+    "node-fetch": "^2.6.1"
   },
   "devDependencies": {
     "@types/jest": "^26.0.14",
     "@types/node-fetch": "^2.5.7",
     "@typescript-eslint/eslint-plugin": "^4.6.0",
     "@typescript-eslint/parser": "^4.6.0",
-    "docdash": "^1.2.0",
     "dotenv": "^8.2.0",
     "eslint": "^7.10.0",
     "eslint-config-prettier": "^6.12.0",
     "eslint-plugin-jest": "^24.0.2",
     "husky": "^4.3.0",
     "jest": "^26.5.0",
-    "jsdoc": "^3.6.6",
     "lint-staged": "^10.4.0",
     "prettier": "^2.1.2",
     "ts-jest": "^26.4.3",
     "ts-loader": "^8.0.12",
     "ts-node": "^9.0.0",
+    "typedoc": "^0.19.2",
     "typescript": "^4.0.5",
     "webpack": "^5.3.2",
     "webpack-cli": "^4.1.0"
