+++ src/asyncAirtable.ts
@@ -21,62 +21,6 @@ declare global {
   }
 }
 
-/**
- * @typedef Options
- * @type {Object}
- * @description An object of possible options. This object cannot be initialized, it is for reference only.
- * @property {string[]} [fields] - An array of specific field names to be returned. Returns all fields if none are supplied.
- * @property {string} [filterByFormula] - A [formula used](https://support.airtable.com/hc/en-us/articles/203255215-Formula-Field-Reference) to filter the records
- * @property {number} [maxRecords=100] - The maximum total number of records that will be returned in your requests. Should be smaller than or equal to `pageSize`.
- * @property {number} [pageSize=100] - The number of records returned in each request. Must be less than or equal to 100.
- * @property {SortOption[]} [sort] - A list of sort objects that specifies how the records will be ordered.
- * @property {string} [view] -
- * The name or id of a view on the specified table.
- * If set, only the records in that view will be returned.
- * The records will be sorted according to the order
- * of the view unless the sort parameter is included,
- * which overrides that order. Fields hidden in this view
- * will be returned in the results. To only return
- * a subset of fields, use the fields parameter.
- * @example
- * * {
- *    fields: ['name', 'email', 'date'],
- *    filterByFormula: "{name} = 'Paul'",
- *    maxRecords: 50,
- *    pageSize: 10,
- *    sort: [
- *      {
- *        field: "name",
- *        direction: "desc"
- *      },
- *      {
- *        field: "date",
- *        direction: "asc"
- *      }
- *    ],
- *    view: 'Grid view'
- * }
- */
-
-/**
- * @typedef SortOption
- * @type {Object}
- * @property {string} field - The field name you want to sort by
- * @property {string} [direction='asc'] - The direction of the sort
- * @example
- * {
- *    field: "name",
- *    direction: "desc"
- * }
- */
-
-/**
- * @typedef DeleteResponse
- * @type {Object}
- * @property {string} [id] - ID of the deleted record
- * @property {boolean} deleted - Status if a record was deleted
- */
-
 const validOptions: string[] = [
   'fields',
   'filterByFormula',
@@ -87,86 +31,37 @@ const validOptions: string[] = [
 ];
 
 /**
- * @typedef Record
- * @type {Object}
- * @name Record
- * @description The record passed into the createRecord and bulkCreate methods
- * @example
- * {
- *   title: "hello",
- *   description: "world"
- * }
- * @property {object} ...fields - Add a separate property for each field
- */
-
-/**
- * @typedef UpdateRecord
- * @type {Object}
- * @name UpdateRecord
- * @description The record passed into the updateRecord and bulkUpdate methods
- * @example
- * {
- *   id: "recABCDEFGHIJK",
- *   fields: {
- *     title: "hello",
- *     description: "world"
- *   }
- * }
- * @property {string} id - Airtable Record ID (only needed for updates)
- * @property {object} fields - Add a separate property for each field
- */
-
-/**
- * @typedef AirtableRecord
- * @type {Object}
- * @name AirtableRecord
- * @description The record returned by AsyncAirtable
- * @example
- * {
- *   id: "recABCDEFGHIJK",
- *   fields: {
- *     title: "hello",
- *     description: "world"
- *   },
- *   createdTime: 'timestamp'
- * }
- * @property {string} id - Airtable Record ID
- * @property {object} fields - Object of fields in the record
- * @property {string} createdTime - Created timestamp
- */
-
-/**
- * @typedef Config
- * @type {Object}
- * @description An optional object used to instatiate AsyncAirtable
- * @example
- * {
- *    "retryOnRateLimit": true,
- *    "maxRetry": 60000,
- *    "retryTimeout": 5000
- * }
- * @property {boolean} [retryOnRateLimit=true] - This decides whether or not the library will handle retrying a request when rate limited
- * @property {number} [maxRetry=60000] - The maxmium amount of time before the library will stop retrying and timeout when rate limited
- * @property {number} [maxRetry=5000] - The starting timeout for the retry. This will get 50% larger with each try until you hit the maxRetry amount
- */
-
-/**
- * The main AsyncAirtable Library
- * @class
+ * AsyncAirtable
  */
 class AsyncAirtable {
+  /**
+   * @default=true
+   * This decides whether or not the library will
+   * handle retrying a request when rate limited
+   */
   retryOnRateLimit: boolean;
+  /**
+   * @default=60000
+   * The maxmium amount of time before the
+   * library will stop retrying and timeout when rate limited
+   */
   maxRetry: number;
+  /**
+   * @default=5000
+   * The starting timeout for the retry. This will get 50%
+   * larger with each try until you hit the maxRetry amount
+   */
   retryTimeout: number;
+  /** The API Key from AirTable */
   apiKey: string;
+  /** The base id from AirTable */
   base: string;
 
   /**
    * Creates a new instance of the AsyncAirtable library.
-   * @constructor
-   * @param {string} apiKey - The API Key from AirTable
-   * @param {string} base - The base id from AirTable
-   * @param {Config} config - The config to use for this instance of AsyncAirtable
+   * @param apiKey The API Key from AirTable
+   * @param base The base id from AirTable
+   * @param config The config to use for this instance of AsyncAirtable
    */
   constructor(apiKey: string, base: string, config?: Config) {
     if (!apiKey) throw new Error('API Key is required.');
@@ -180,11 +75,10 @@ class AsyncAirtable {
 
   /**
    * Select record(s) from the specified table.
-   * @method
-   * @param {string} table - Table name
-   * @param {Options} [options] - Options object, used to filter records
-   * @param {number} [page] - Used to get a specific page of records
-   * @returns {Promise<AirtableRecord[]>}
+   * @param table Table name
+   * @param options Options object, used to filter records
+   * @param page Used to get a specific page of records
+   * @returns
    */
   select = async (
     table: string,
@@ -287,10 +181,9 @@ class AsyncAirtable {
 
   /**
    * Finds a record on the specified table.
-   * @method
-   * @param {string} table - Table name
-   * @param {string} id - Airtable record ID
-   * @returns {Promise<AirtableRecord>}
+   * @param table Table name
+   * @param id Airtable record ID
+   * @returns
    */
   find = async (table: string, id: string): Promise<AirtableRecord> => {
     try {
@@ -325,10 +218,9 @@ class AsyncAirtable {
 
   /**
    * Creates a new record on the specified table.
-   * @method
-   * @param {string} table - Table name
-   * @param {Record} record - Record object, used to structure data for insert
-   * @returns {Promise<AirtableRecord>}
+   * @param table - Table name
+   * @param record - Record object, used to structure data for insert
+   * @returns
    */
   createRecord = async (
     table: string,
@@ -377,11 +269,10 @@ class AsyncAirtable {
 
   /**
    * Updates a record on the specified table.
-   * @method
-   * @param {string} table - Table name
-   * @param {UpdateRecord} record - Record object, used to update data within a specific record
-   * @param {boolean} [destructive=false] - (Dis-)Allow a destructive update
-   * @returns {Promise<AirtableRecord>}
+   * @param table - Table name
+   * @param record - Record object, used to update data within a specific record
+   * @param destructive - (Dis-)Allow a destructive update
+   * @returns
    */
   updateRecord = async (
     table: string,
@@ -431,10 +322,9 @@ class AsyncAirtable {
 
   /**
    * Deletes a record from the specified table.
-   * @method
-   * @param {string} table - Table name
-   * @param {string} id - Airtable record ID
-   * @returns {Promise<DeleteResponse>}
+   * @param table - Table name
+   * @param id - Airtable record ID
+   * @returns
    */
   deleteRecord = async (table: string, id: string): Promise<DeleteResponse> => {
     try {
@@ -475,10 +365,9 @@ class AsyncAirtable {
 
   /**
    * Creates multiple new records on the specified table.
-   * @method
-   * @param {string} table - Table name
-   * @param {Array<Record>} records - An array of Record objects
-   * @returns {Promise<AirtableRecord[]>}
+   * @param table - Table name
+   * @param records - An array of Record objects
+   * @returns
    */
   bulkCreate = async (
     table: string,
@@ -530,10 +419,9 @@ class AsyncAirtable {
 
   /**
    * Updates multiple records on the specified table
-   * @method
-   * @param {string} table - Table name
-   * @param {Array<UpdateRecord>} records - An array of Record objects
-   * @returns {Promise<AirtableRecord[]>}
+   * @param table - Table name
+   * @param records - An array of Record objects
+   * @returns
    */
   bulkUpdate = async (
     table: string,
@@ -582,10 +470,9 @@ class AsyncAirtable {
 
   /**
    * Deletes multiple records from the specified table
-   * @method
-   * @param {string} table - Table name
-   * @param {Array<string>} ids - Array of Airtable record IDs
-   * @returns {Promise<DeleteResponse[]>}
+   * @param table - Table name
+   * @param ids - Array of Airtable record IDs
+   * @returns
    */
   bulkDelete = async (
     table: string,
@@ -638,11 +525,10 @@ class AsyncAirtable {
 
   /**
    * Checks if a record exists, and if it does updates it, if not creates a new record.
-   * @method
-   * @param {string} table - Table name
-   * @param {string} filterString - The filter formula string used to check for a record
-   * @param {Record} record - Record object used to either update or create a record
-   * @returns {Promise<AirtableRecord>}
+   * @param table - Table name
+   * @param filterString - The filter formula string used to check for a record
+   * @param record - Record object used to either update or create a record
+   * @returns
    */
   upsertRecord = async (
     table: string,
